+++ app/src/main/java/fi/tv7/taivastv7/fragments/ProgramInfoFragment.java
@@ -32,6 +32,7 @@ import fi.tv7.taivastv7.helpers.Sidebar;
 import fi.tv7.taivastv7.helpers.Utils;
 import fi.tv7.taivastv7.model.SharedCacheViewModel;
 
+import static fi.tv7.taivastv7.helpers.Constants.ARCHIVE_MAIN_FRAGMENT;
 import static fi.tv7.taivastv7.helpers.Constants.ARCHIVE_PLAYER_FRAGMENT;
 import static fi.tv7.taivastv7.helpers.Constants.ASPECT_RATIO;
 import static fi.tv7.taivastv7.helpers.Constants.ASPECT_RATIO_16_9;
@@ -426,9 +427,11 @@ public class ProgramInfoFragment extends Fragment {
                 }
                 else {
                     String toPage = sharedCacheViewModel.getPageFromHistory();
-                    if (toPage != null) {
-                        Utils.toPage(toPage, getActivity(), true, false,null);
+                    if (toPage == null) {
+                        toPage = ARCHIVE_MAIN_FRAGMENT;
                     }
+
+                    Utils.toPage(toPage, getActivity(), true, false,null);
                 }
             }
         }