+++ app/src/main/java/fi/tv7/taivastv7/helpers/Constants.java
@@ -180,7 +180,7 @@ abstract public class Constants {
     public static final String QUERY_PARAM = "query";
 
     public static final int RECOMMENDED_PROGRAMS_LIMIT = 30;
-    public static final int NEWEST_LIMIT = 20;
+    public static final int NEWEST_LIMIT = 30;
     public static final int CATEGORY_PROGRAMS_SEARCH_LIMIT = 30;
     public static final int SERIES_PROGRAMS_SEARCH_LIMIT = 30;